+++ e2e/testsuite/codec.go
@@ -25,7 +25,6 @@ import (
 	gmptypes "github.com/cosmos/ibc-go/v11/modules/apps/27-gmp/types"
 	icacontrollertypes "github.com/cosmos/ibc-go/v11/modules/apps/27-interchain-accounts/controller/types"
 	icahosttypes "github.com/cosmos/ibc-go/v11/modules/apps/27-interchain-accounts/host/types"
-	ratelimitingtypes "github.com/cosmos/ibc-go/v11/modules/apps/rate-limiting/types"
 	transfertypes "github.com/cosmos/ibc-go/v11/modules/apps/transfer/types"
 	v7migrations "github.com/cosmos/ibc-go/v11/modules/core/02-client/migrations/v7"
 	clienttypes "github.com/cosmos/ibc-go/v11/modules/core/02-client/types"
@@ -76,7 +75,6 @@ func codecAndEncodingConfig() (*codec.ProtoCodec, testutil.TestEncodingConfig) {
 	wasmtypes.RegisterInterfaces(cfg.InterfaceRegistry)
 	attestationstypes.RegisterInterfaces(cfg.InterfaceRegistry)
 	channeltypesv2.RegisterInterfaces(cfg.InterfaceRegistry)
-	ratelimitingtypes.RegisterInterfaces(cfg.InterfaceRegistry)
 
 	// all other types
 	upgradetypes.RegisterInterfaces(cfg.InterfaceRegistry)