+++ modules/apps/transfer/keeper/keeper_test.go
@@ -17,7 +17,6 @@ import (
 	authkeeper "github.com/cosmos/cosmos-sdk/x/auth/keeper"
 	minttypes "github.com/cosmos/cosmos-sdk/x/mint/types"
 
-	ratelimiting "github.com/cosmos/ibc-go/v11/modules/apps/rate-limiting"
 	"github.com/cosmos/ibc-go/v11/modules/apps/transfer/keeper"
 	"github.com/cosmos/ibc-go/v11/modules/apps/transfer/types"
 	ibctesting "github.com/cosmos/ibc-go/v11/testing"
@@ -334,16 +333,9 @@ func (s *KeeperTestSuite) TestUnsetParams() {
 func (s *KeeperTestSuite) TestWithICS4Wrapper() {
 	s.SetupTest()
 
-	// test if the ics4 wrapper is the rate limit keeper initially
-	ics4Wrapper := s.chainA.GetSimApp().TransferKeeper.GetICS4Wrapper()
-
-	_, isRatelimitingKeeper := ics4Wrapper.(*ratelimiting.IBCMiddleware)
-	s.Require().True(isRatelimitingKeeper)
-	s.Require().IsType((*ratelimiting.IBCMiddleware)(nil), ics4Wrapper)
-
 	// set the ics4 wrapper to the channel keeper
 	s.chainA.GetSimApp().TransferKeeper.WithICS4Wrapper(nil)
-	ics4Wrapper = s.chainA.GetSimApp().TransferKeeper.GetICS4Wrapper()
+	ics4Wrapper := s.chainA.GetSimApp().TransferKeeper.GetICS4Wrapper()
 	s.Require().Nil(ics4Wrapper)
 }