+++ CHANGELOG.md
@@ -38,7 +38,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### Features
 
-* [\#8573](https://github.com/cosmos/ibc-go/pull/8573) Support custom address codecs in transfer and rate limiting.
+* [\#8573](https://github.com/cosmos/ibc-go/pull/8573) Support custom address codecs in transfer.
 * [\#8285](https://github.com/cosmos/ibc-go/pull/8285) Packet forward middleware.
 * [\#8545](https://github.com/cosmos/ibc-go/pull/8545) Support sending multiple payloads in the same packet for atomic payload execution.
 * [\#8473](https://github.com/cosmos/ibc-go/pull/8473) Support sending v2 packets on v1 channel identifiers using aliasing.
@@ -65,8 +65,6 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Improvements
 
 ### Bug Fixes
-
-(apps/rate-limiting) [\#8767](https://github.com/cosmos/ibc-go/pull/8767) Fix string conflict in rate-limiting prefix iterator
 
 ### Testing API